Dockerfile: add path containing cyclonedx to PATH
Allows the cyclonedx command to be more easily used when running the image with an entrypoint.

For example, when running:
```
docker run -w /app -v "$(pwd)":/app:Z -it --entrypoint /bin/bash cyclonedx/cyclonedx-cli:0.24.0
```
it's nice to be able to use the "cyclonedx" command instead of having to give an absolute path ("/cyclonedx").

Signed-off-by: Craig Andrews <candrews@integralblue.com>