Removed trimming of binaries
The CLI was unable to sign XML BOMs as can be seen in: https://github.com/CycloneDX/cyclonedx-cli/issues/251

This is caused by the trimming of binaries; when publishing the binary, the build will cause several IL2026 warnings in the crypto library and xml library. According to ,https://learn.microsoft.com/en-us/dotnet/core/deploying/trimming/fixing-warnings , this warning denotes: "This attribute is used when code is fundamentally not trim compatible, or the trim dependency is too complex to explain to the trimmer. "

More information can be seen here: https://learn.microsoft.com/en-us/dotnet/core/deploying/trimming/trim-warnings/il2026

Therefore, I think that trimming should be disabled to ensure proper working of the tool.

Signed-off-by: Martin van Hensbergen <mhens@mac.com>