Overwrite existing files when a new SBOM is saved
A change is made so when a new SBOM is produced it overwrites any
existing file with the same name. This change is introduced due to the
current behavior which simply writes the bytes to the same file, so if
the previous files length was longer, that content would still remain.
Leading to an invalid SBOM being created by the CLI.

Signed-off-by: Tristan Farkas <Tristan.Farkas@axis.com>